Create draft for loopback spike
Ignore the deletion of SW_UART.ino I just compiled it and Arduino IDE
automatically creates a folder.

Main reason for this pull request is a draft for the loopback mode (task
in sprint 3):
Spike compiles but does not do what it is supposed to do for some
reason. Goal was to receive a character and send it back immediately.
Currently it is setup so it as soon as a character is send FROM the
serial monitor, it SHOULD print out A debug line within the serial
monitor but also send it out to the wifly modul and it should be
displayed at whatever telnet communication tool is used.
What acutally happens now is that the debug line is printed out a
gazzilion times and nothing is send to the Wifly modul. I also tried
other option and when I send characters/words through the Serial monitor
they arrive just fine at the Wifly modul, but through a variable it is
not working.