Add timer 1 example
This is to show the embedded setup, triggering and use in a loop for
timer 1.

Use:
Timer 1 is setup in function 'SetupTimer1' for 1 second intervals.
Everytime the timer1 is triggerd variable 'g_traffic_light_counter' is
counted up. In the Main loop global variable 'g_traffic_light_counter'
is processed.

All timer related code is marked with 'TIMER_EXAMPLE' for your
convenience.