Include header files into Arduino IDE
First I create the .h  in Notepad.  This works just fines and compiles
but has the drawack that compiler errors do not show up IN .h those
file.

In this commit I created the .h in the Arduino IDE. Errors will be shown
in the IDE now. More convenient for debugging. Drawback is that I have
to move the .h files from the include folder into the main directory
where the main.ino resides.