Add Iza's timer to integration, change folder structure for libraries
External Libraries must be in the same folder as the main sketch if they
are not in the local Arduino Library directory. The latter is not
desired since we want to keep everythin in one place. (Thus notice that
Arduino.h was deleted from library folder and inserted a directory above
it)