Extension timer
New feature: the remaining time the factory needs to end all orders can
be requested and it will be communicated through serial monitor. The
request is simulated with a random number generator: each time the
number 2 is picked, the uC sends to the serial monitor the time info,
that the factory still needs for processing the orders. Please merge if
you agree.