Add most recent timer code to integration and delete old code
Reason for abolishing all changes previously made is that I will take
the code that was tested by Iza in spike 'Timer_example2' and transfer
it with minimal changes to the integration folder. Thus avoiding bugs.

In this push the driver_timer files were updated and the driver_timer.h
was included in the Main.c file. I then tested the bi-directional
communication between App and Arduino and it stillt works.

In futur commits I will implement the timer functionality in the Main.c
file.

Merging is not necessary at this point.