Add timer timer example functionality to Main.ino file
This example will print the current time of the three (3) timers.
I tested it and all data is printed to the Serial Monitor. Communication
to App still works (assigning leasen *OPEN*, *CLOS*, normal messages). I
could not communicate from app to Serial Monitor due to the issues I
have with Android 7. I'll attach a screenshot of the Serial Monitor
output.